implemented device_read

+++ message_slot.h
@@ -14,8 +14,8 @@
 typedef struct channel_node
 {
     struct channel_node next;
-    int msg_len; // actual length of message - bounded by 128 bytes
-    char msg[BUF_LEN]; // the message
+    size_t msg_len; // actual length of message - bounded by 128 bytes
+    char msg[BUF_LEN]; // the message. Note that the size is already set at instantiation (it is O(1))
     unsigned long id; // this not the message_slot minor, so is not bounded in size
 } ch_node ;
